Move TestOpenAPIDelegationChainPlumbing to its own file
